Added code gen structure
