+++ src/codegen/codegen.h
@@ -1,0 +1,11 @@
+/* Represents the type of instruction */
+enum InstructionType
+{
+	PUSH_I, PUSH_R
+};
+
+/* Represents an instruction */
+struct Instruction
+{
+	enum InstructionType type;
+};
