+++ src/codegen/registers.h
@@ -1,0 +1,5 @@
+/* All x86 registers */
+enum Register
+{
+	EAX, EBX, ECX, EDX
+};
